정규식 관련 함수를 위한 구조체 커터 생성 및 파일생성 정규식화 해주는 함수 makeRegexFrom 작성
