+++ CoordinateCalculator/CoordinateCalculator/OutputView.swift
@@ -2,13 +2,13 @@
 //  OutputView.swift
 //  CoordinateCalculator
 //
-//  Created by Yoda Codd on 2018. 4. 9..
+//  Created by Yoda Codd on 2018. 4. 10..
 //  Copyright © 2018년 Codesquad Inc. All rights reserved.
 //
 
 import Foundation
 
-func drawAxis(){
-//    let ansiDrawPattern = "\u{001B}["
-    
+struct OutputView(){
+    func drawAxis(){
+    }
 }
